requirements.txt
transformers>=4.36.0
accelerate>=0.25.0
datasets>=2.14.0
bitsandbytes>=0.41.0
peft>=0.5.0
llama-cpp-python>=0.2.0
jupyter
gradio>=3.50.0
wandb>=0.15.0
hydra-core>=1.3.0
omegaconf>=2.3.0
pytorch-lightning>=2.0.0
torch>=2.0.0
flash-attn>=2.3.0
einops
scipy
tqdm 